Vendor-Upgrade v4.3.3

+++ typo3/sysext/indexed_search/class.indexer.php
@@ -624,8 +624,12 @@ class tx_indexedsearch_indexer {
 			for($i=0;$this->embracingTags($headPart,'meta',$dummy,$headPart,$meta[$i]);$i++) { /*nothing*/ }
 			for($i=0;isset($meta[$i]);$i++) {
 				$meta[$i] = t3lib_div::get_tag_attributes($meta[$i]);
-				if(stristr($meta[$i]['name'],'keywords')) $contentArr['keywords'].=','.$meta[$i]['content'];
-				if(stristr($meta[$i]['name'],'description')) $contentArr['description'].=','.$meta[$i]['content'];
+				if (stristr($meta[$i]['name'], 'keywords')) {
+					$contentArr['keywords'] .= ',' . $this->addSpacesToKeywordList($meta[$i]['content']);
+				}
+				if (stristr($meta[$i]['name'], 'description')) {
+					$contentArr['description'] .= ',' . $meta[$i]['content'];
+				}
 			}
 		}
 
@@ -790,6 +794,7 @@ class tx_indexedsearch_indexer {
 					$this->indexExternalUrl($linkSource);
 				}
 			} elseif (!$qParts['query']) {
+				$linkSource = urldecode($linkSource);
 				if (t3lib_div::isAllowedAbsPath($linkSource))	{
 					$localFile = $linkSource;
 				} else {
@@ -913,12 +918,14 @@ class tx_indexedsearch_indexer {
 			if (strlen($content))	{
 
 					// Create temporary file:
-				$tmpFile = t3lib_div::tempnam('EXTERNAL_URL').'.html';
-				t3lib_div::writeFile($tmpFile, $content);
+				$tmpFile = t3lib_div::tempnam('EXTERNAL_URL');
+				if ($tmpFile) {
+					t3lib_div::writeFile($tmpFile, $content);
 
-					// Index that file:
-				$this->indexRegularDocument($externalUrl, TRUE, $tmpFile, 'html');	// Using "TRUE" for second parameter to force indexing of external URLs (mtime doesn't make sense, does it?)
-				unlink($tmpFile);
+						// Index that file:
+					$this->indexRegularDocument($externalUrl, TRUE, $tmpFile, 'html');	// Using "TRUE" for second parameter to force indexing of external URLs (mtime doesn't make sense, does it?)
+					unlink($tmpFile);
+				}
 			}
 		}
 	}
@@ -2084,6 +2091,19 @@ class tx_indexedsearch_indexer {
 
 		require_once t3lib_extMgm::extPath('indexed_search') . 'hooks/class.tx_indexedsearch_tslib_fe_hook.php';
 		t3lib_div::makeInstance('tx_indexedsearch_tslib_fe_hook')->headerNoCache($params, $ref);
+	}
+
+	/**
+	 * Makes sure that keywords are space-separated. This is impotant for their
+	 * proper displaying as a part of fulltext index.
+	 *
+	 * @param string $keywordList
+	 * @return string
+	 * @see http://bugs.typo3.org/view.php?id=1436
+	 */
+	protected function addSpacesToKeywordList($keywordList) {
+		$keywords = t3lib_div::trimExplode(',', $keywordList);
+		return ' ' . implode(', ', $keywords) . ' ';
 	}
 }
